requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
python-multipart==0.0.9
jinja2==3.1.4
Pillow==9.5.0
pillow-heif==0.10.0
ffmpeg-python==0.2.0